\name{demo.dag2}
\Rdversion{1.1}
\alias{demo.dag2}
\title{
Set up demo DAG #2.
}
\description{
Initializes a more complex DAG, motivated by Shrier (2008).
 This DAG was used to examine the performance of \code{brute.search} and
 has been featured in Breitling (in press).
}
\usage{
demo.dag2()
}
\details{
}
\value{
Returns a DAG.
}
\references{
Breitling LP (in press). dagR: a suite of R functions for directed acyclic graphs. Epidemiology.\cr
Shrier I, Platt RW (2008). Reducing bias through directed acyclic graphs. BMC Med Res Methodol 8:70
}
\author{
Lutz P Breitling <lutz.breitling@gmail.com>
}
\note{
}


\seealso{
\code{demo.dag0}, \code{demo.dag1}, \code{demo.dag3},
 \code{demo.dag4}, \code{demo.dag5}, \code{demo.dag6}
}
\examples{
demo.dag2();
}
