\name{demo.dag3}
\Rdversion{1.1}
\alias{demo.dag3}
\encoding{UTF-8}
\title{
Set up demo DAG #3.
}
\description{
Initializes a DAG motivated by the manual for the software DAG v0.11
 (Knüppel 2009). This DAG has been featured in Breitling (2010).
}
\usage{
demo.dag3()
}
\value{
Returns a DAG.
}
\references{
Breitling LP (2010). dagR: a suite of R functions for directed acyclic graphs. Epidemiology 21(4):586-587.\cr
Knüppel S (2009). DAG v0.11 documentation (Oct 21, 2009). \url{https://hsz.dife.de/dag/}
}
\author{
Lutz P Breitling <l.breitling@posteo.de>
}


\seealso{
\code{demo.dag0}, \code{demo.dag1}, \code{demo.dag2},
 \code{demo.dag4}, \code{demo.dag5}, \code{demo.dag6}
}

