% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_safety_copy.R
\name{make_safety_copy}
\alias{make_safety_copy}
\title{Make a safety copy of a file}
\usage{
make_safety_copy(fn, length_file_extension = 3)
}
\arguments{
\item{fn}{The file name}

\item{length_file_extension}{Length of the file extension, by default 3 (for
instance .png, .jpg, .txt, .csv). If the file extension is shorter or longer
(e.g. .html), specify that here. A couple of simple tests is done to make sure
that length_file_extension has been set correctly.}
}
\value{
Nothing (NULL)
}
\description{
Makes a safety copy of a file with "_OLD" and a number in the file name, before you are
overwriting an existing file. For instance, say that you have made a file
called 'my_precious_figure.png' in the 'figures' directory. If you call
\code{make_safety_copy("figures/my_precious_figure.png")}, a safety copy named
\code{'my_precious_figure_OLD1.png'} will be made. However, if the 'OLD01' name
already has been used, the copy will be named 'my_precious_figure_OLD2.png'. If that
name also has been used, 'my_precious_figure_OLD3.png' will be used, etcetera.
}
\examples{

# Will make a copy named 'Figures/07_tbt_vs_year_FIG7A_OLD01.png'
fn <- "Figures/07_tbt_vs_year_FIG7A.png"
make_safety_copy(fn)
# (if you run it again, the copy will be called
'Figures/07_tbt_vs_year_FIG7A_OLD02.png')

Too long or short file extension will throw a warning (and make a file copy with
a strange name):
make_safety_copy("Figures/My_test_file.html")
make_safety_copy("Figures/My_test_file.tx")

# Adjust length_file_extension to get a copy with a better name:
make_safety_copy("Figures/My_test_file.html", length_file_extension = 4)
make_safety_copy("Figures/My_test_file.tx", length_file_extension = 2)
}
