\name{print.dagRdag}
\alias{print.dagRdag}
\title{
Prints the raw contents of an object of class dagRdag.
}
\description{
Generic function \code{print} code for class \code{dagRdag}. This uses the default \code{print} method for list objects and points the user to the availability of the more convenient \code{summary} method.
}
\usage{
\method{print}{dagRdag}(x, ...)
}
\arguments{
  \item{x}{
An object of class \code{dagRdag}.}
  \item{...}{
Other arguments passed to the print routine.
}
}

\author{
Lutz P Breitling <l.breitling@posteo.de>
}
