\name{demo.dag0}
\Rdversion{1.1}
\alias{demo.dag0}
\title{
Set up demo DAG #0.
}
\description{
Initializes a simple DAG used during the dagR development phase.
}
\usage{
demo.dag0()
}
\value{
Returns a DAG.
}
\author{
Lutz P Breitling <l.breitling@posteo.de>
}


\seealso{
\code{demo.dag1}, \code{demo.dag2}, \code{demo.dag3},
 \code{demo.dag4}, \code{demo.dag5}, \code{demo.dag6}
}
