\name{demo.dag1}
\Rdversion{1.1}
\alias{demo.dag1}
\title{
Set up demo DAG #1.
}
\description{
Initializes a classical "M DAG" useful for demonstrating harmful adjustment.
 The DAG is motivated by figure 3 in Fleischer (2008)
 and also featured in Breitling (in press).
}
\usage{
demo.dag1()
}
\details{
}
\value{
Returns a DAG.
}
\references{
Breitling LP (in press). dagR: a suite of R functions for directed acyclic graphs. Epidemiology.\cr
Fleischer NL, Diez Roux AV (2008). Using directed acyclic graphs to guide analyses of neighbourhood health effects: an introduction. J Epidemiol Community Health 62:842-846.
}
\author{
Lutz P Breitling <lutz.breitling@gmail.com>
}
\note{
}


\seealso{
\code{demo.dag0}, \code{demo.dag2}, \code{demo.dag3},
 \code{demo.dag4}, \code{demo.dag5}, \code{demo.dag6}
}
\examples{
demo.dag1();
}
