\name{demo.dag3}
\Rdversion{1.1}
\alias{demo.dag3}
\title{
Set up demo DAG #3.
}
\description{
Initializes a DAG motivated by the manual for the software DAG v0.11
 (Kn\"uppel 2009). This DAG has been featured in Breitling (in press).
}
\usage{
demo.dag3()
}
\details{
}
\value{
Returns a DAG.
}
\references{
Breitling LP (in press). dagR: a suite of R functions for directed acyclic graphs. Epidemiology.\cr
Kn\"uppel S (2009). DAG v0.11 documentation (Oct 21, 2009). \url{http://epi.dife.de/dag/}
}
\author{
Lutz P Breitling <lutz.breitling@gmail.com>
}
\note{
}


\seealso{
\code{demo.dag0}, \code{demo.dag1}, \code{demo.dag2},
 \code{demo.dag4}, \code{demo.dag5}, \code{demo.dag6}
}
\examples{
demo.dag3();
}
